backend/requirements.txt
# FastAPI Backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Core dependencies (from existing project)
numpy>=1.24.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
scipy>=1.11.0
scikit-image>=0.21.0
pycryptodome>=3.18.0
ffmpeg-python>=0.2.0

# CORS
fastapi-cors>=0.0.6

# File handling
aiofiles>=24.1.0
